requirements.txt
psutil==5.9.4
frida-tools==10.2.0
